+++ src/Autocomplete/stories/overview.jsx
@@ -1,0 +1,6 @@
+import React from 'react'
+
+import ComponentOverview from '../../ComponentOverview'
+import { Autocomplete } from '../Autocomplete'
+
+export default () => <ComponentOverview component={Autocomplete} />