+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 * Upgrade storybook to 5.0.5
+* Add `Autocomplete` component
 
 ## 1.14.0 (2019-03-26)